--- a/src/arith_uint256.cpp
+++ b/src/arith_uint256.cpp
@@ -8,8 +8,7 @@
 #include <uint256.h>
 #include <crypto/common.h>
 
-#include <stdio.h>
-#include <string.h>
+#include <cassert>
 
 template <unsigned int BITS>
 base_uint<BITS>::base_uint(const std::string& str)