--- a/src/arith_uint256.h
+++ b/src/arith_uint256.h
@@ -6,7 +6,6 @@
 #ifndef BITCOIN_ARITH_UINT256_H
 #define BITCOIN_ARITH_UINT256_H
 
-#include <assert.h>
 #include <cstring>
 #include <limits>
 #include <stdexcept>