--- a/src/arith_uint256.h
+++ b/src/arith_uint256.h
@@ -12,7 +12,6 @@
 #include <stdexcept>
 #include <stdint.h>
 #include <string>
-#include <vector>
 
 class uint256;