--- a/src/bitcoin-cli.cpp
+++ b/src/bitcoin-cli.cpp
@@ -14,8 +14,10 @@
 #include <rpc/protocol.h>
 #include <rpc/request.h>
 #include <util/strencodings.h>
+
 #include <util/system.h>
 #include <util/translation.h>
+#include <util/time.h>
 
 #include <functional>
 #include <memory>