--- a/src/coins.h
+++ b/src/coins.h
@@ -8,7 +8,6 @@
 
 #include <primitives/transaction.h>
 #include <compressor.h>
-#include <core_memusage.h>
 #include <crypto/siphash.h>
 #include <memusage.h>
 #include <serialize.h>