--- a/src/consensus/merkle.h
+++ b/src/consensus/merkle.h
@@ -7,7 +7,6 @@
 
 #include <vector>
 
-#include <primitives/transaction.h>
 #include <primitives/block.h>
 #include <uint256.h>