--- a/src/crypto/aes.cpp
+++ b/src/crypto/aes.cpp
@@ -4,7 +4,6 @@
 
 #include <crypto/aes.h>
 
-#include <assert.h>
 #include <string.h>
 
 extern "C" {