--- a/src/httprpc.cpp
+++ b/src/httprpc.cpp
@@ -7,13 +7,14 @@
 #include <chainparams.h>
 #include <crypto/hmac_sha256.h>
 #include <httpserver.h>
-#include <key_io.h>
 #include <rpc/protocol.h>
 #include <rpc/server.h>
-#include <sync.h>
-#include <ui_interface.h>
-#include <util/strencodings.h>
+
+#include <util/memory.h>
 #include <util/system.h>
+#include <util/strencodings.h>
+#include <util/time.h>
+#include <ui_interface.h>
 #include <util/translation.h>
 #include <walletinitinterface.h>