--- a/src/httprpc.h
+++ b/src/httprpc.h
@@ -5,9 +5,6 @@
 #ifndef BITCOIN_HTTPRPC_H
 #define BITCOIN_HTTPRPC_H
 
-#include <string>
-#include <map>
-
 /** Start HTTP RPC subsystem.
  * Precondition; HTTP and RPC has been started.
  */