--- a/src/index/txindex.cpp
+++ b/src/index/txindex.cpp
@@ -4,7 +4,9 @@
 
 #include <index/txindex.h>
 #include <shutdown.h>
+#include <txdb.h>
 #include <ui_interface.h>
+#include <util/memory.h>
 #include <util/system.h>
 #include <util/translation.h>
 #include <validation.h>