--- a/src/index/txindex.h
+++ b/src/index/txindex.h
@@ -7,7 +7,6 @@
 
 #include <chain.h>
 #include <index/base.h>
-#include <txdb.h>
 
 /**
  * TxIndex is used to look up transactions included in the blockchain by hash.