--- a/src/noui.cpp
+++ b/src/noui.cpp
@@ -3,10 +3,10 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include <logging.h>
 #include <noui.h>
 
 #include <ui_interface.h>
-#include <util/system.h>
 
 #include <string>