--- a/src/policy/fees.cpp
+++ b/src/policy/fees.cpp
@@ -6,10 +6,10 @@
 #include <policy/fees.h>
 
 #include <clientversion.h>
-#include <primitives/transaction.h>
 #include <streams.h>
 #include <txmempool.h>
 #include <util/system.h>
+#include <util/time.h>
 
 static constexpr double INF_FEERATE = 1e99;