--- a/src/psbt.cpp
+++ b/src/psbt.cpp
@@ -5,8 +5,6 @@
 #include <psbt.h>
 #include <util/strencodings.h>
 
-#include <numeric>
-
 PartiallySignedTransaction::PartiallySignedTransaction(const CMutableTransaction& tx) : tx(tx)
 {
     inputs.resize(tx.vin.size());