--- a/src/qt/bantablemodel.cpp
+++ b/src/qt/bantablemodel.cpp
@@ -7,8 +7,6 @@
 #include <qt/clientmodel.h>
 
 #include <interfaces/node.h>
-#include <sync.h>
-#include <util/time.h>
 
 #include <QDebug>
 #include <QList>