--- a/src/qt/bitcoin.cpp
+++ b/src/qt/bitcoin.cpp
@@ -38,8 +38,6 @@
 
 #include <memory>
 
-#include <boost/thread.hpp>
-
 #include <QApplication>
 #include <QDebug>
 #include <QLibraryInfo>