--- a/src/qt/sendcoinsdialog.cpp
+++ b/src/qt/sendcoinsdialog.cpp
@@ -23,9 +23,8 @@
 #include <key_io.h>
 #include <wallet/coincontrol.h>
 #include <ui_interface.h>
-#include <txmempool.h>
 #include <policy/fees.h>
-#include <wallet/fees.h>
+#include <wallet/wallet.h>
 
 #include <QFontMetrics>
 #include <QScrollBar>