--- a/src/qt/walletmodel.h
+++ b/src/qt/walletmodel.h
@@ -7,8 +7,8 @@
 
 #include <amount.h>
 #include <key.h>
+#include <primitives/transaction.h>
 #include <serialize.h>
-#include <script/standard.h>
 
 #if defined(HAVE_CONFIG_H)
 #include <config/bitcoin-config.h>