--- a/src/qt/walletmodeltransaction.h
+++ b/src/qt/walletmodeltransaction.h
@@ -7,7 +7,6 @@
 
 #include <qt/walletmodel.h>
 
-#include <memory>
 #include <amount.h>
 
 #include <QObject>