--- a/src/qt/winshutdownmonitor.cpp
+++ b/src/qt/winshutdownmonitor.cpp
@@ -6,7 +6,6 @@
 
 #if defined(Q_OS_WIN)
 #include <shutdown.h>
-#include <util/system.h>
 
 #include <windows.h>