--- a/src/rpc/blockchain.cpp
+++ b/src/rpc/blockchain.cpp
@@ -14,6 +14,7 @@
 #include <core_io.h>
 #include <hash.h>
 #include <index/blockfilterindex.h>
+#include <key.h>
 #include <policy/feerate.h>
 #include <policy/policy.h>
 #include <policy/rbf.h>