--- a/src/rpc/rawtransaction.cpp
+++ b/src/rpc/rawtransaction.cpp
@@ -32,7 +32,6 @@
 #include <validationinterface.h>
 
 
-#include <numeric>
 #include <stdint.h>
 
 #include <univalue.h>