--- a/src/rpc/rawtransaction.cpp
+++ b/src/rpc/rawtransaction.cpp
@@ -6,7 +6,8 @@
 #include <chain.h>
 #include <coins.h>
 #include <consensus/validation.h>
-#include <core_io.h>
+#include <core_read.h>
+#include <core_write.h>
 #include <index/txindex.h>
 #include <key_io.h>
 #include <merkleblock.h>