--- a/src/rpc/server.h
+++ b/src/rpc/server.h
@@ -8,7 +8,6 @@
 
 #include <amount.h>
 #include <rpc/request.h>
-#include <uint256.h>
 
 #include <map>
 #include <stdint.h>