--- a/src/rpc/util.h
+++ b/src/rpc/util.h
@@ -10,7 +10,6 @@
 #include <rpc/protocol.h>
 #include <rpc/request.h>
 #include <script/script.h>
-#include <script/sign.h>
 #include <script/standard.h>
 #include <univalue.h>
 #include <util/error.h>