--- a/src/script/standard.h
+++ b/src/script/standard.h
@@ -11,8 +11,6 @@
 
 #include <boost/variant.hpp>
 
-#include <stdint.h>
-
 static const bool DEFAULT_ACCEPT_DATACARRIER = true;
 
 class CKeyID;