--- a/src/support/events.h
+++ b/src/support/events.h
@@ -5,8 +5,8 @@
 #ifndef BITCOIN_SUPPORT_EVENTS_H
 #define BITCOIN_SUPPORT_EVENTS_H
 
-#include <ios>
 #include <memory>
+#include <string>
 
 #include <event2/event.h>
 #include <event2/http.h>