--- a/src/sync.cpp
+++ b/src/sync.cpp
@@ -16,7 +16,6 @@
 #include <stdio.h>
 
 #include <map>
-#include <memory>
 #include <set>
 
 #ifdef DEBUG_LOCKCONTENTION