@hebasto 2019-06-24
Just wondering what is a proper way to test such PRs?