@MarcoFalke 2019-06-24
Could create a separate pull for all changes in  `src/bench/ src/test src/wallet/test/`?