@practicalswift 2019-06-24
@MarcoFalke Good idea! Submitted as #16278 :-)