@Sjors 2019-06-27
Good idea to move the tests to #16027. The commit `splitting core_io.h into the ex…
…pected core_read.h/core_write.h` could also be a separate PR if this ends up taking longer.

This touches quite a few other PRs, but I suspect most won't trigger a rebase. Otherwise it would be best to merge this right after a major release (@MarcoFalke would be cool if @DrahtBot could tell). So concept ACK.

Suggest moving "Reduce memory usage during build by" from the commit name(s) to the commit message(s); emphasize _what_ you're changing in the title, and _why_ in the message.

I was able to run unit and functional* tests on macOS with 67fc141 (rebased on on master).

* except `p2p_invalid_messages.py` as usual