@practicalswift 2019-06-27
@Sjors Thanks for taking the time to review and test!

Good point about the commit messages! I've now changed that. I guess I was eager to explain the *why* to also reach those who form their judgement based on PR metadata alone: the GitHub equivalent of having an opinion about the content of a news articles after only reading the title :-)