@practicalswift 2019-07-02
> Just wondering what is a proper way to test such PRs?

Making sure we compile and pass the test suite (both functional and unit tests) on supported platforms would be the obvious starting point I guess? :-)