@practicalswift 2019-07-03
Rebased! :-)