@practicalswift 2019-07-10
Rebased! :-)