@practicalswift 2019-07-21
Rebased! :-)