@MarcoFalke 2019-08-14
@practicalswift While this is probably nothing to be merged as-is, I still liked having the pull request open. You might have noticed that I fixed the includes in files that I touched based on the diff in this pull. Would you mind re-opening and/or sharing the steps you took to generate this list?

I tried to run iwyu once, but I wasn't really successful. So if your solution works better, why not put a comment in #15442?